Refactor AssetService and TagsController for asset management and tagging functionality. Merge with resizing branch.
