+++ DAM.Backend/Controllers/AssetsController/TagsController.cs
@@ -19,10 +19,10 @@ public class TagsController : ApiController
      * Gets all tags
      */
     [HttpGet()]
-    public async Task<IActionResult> GetAllTags()
+    public async Task<IActionResult> GetTags()
     {
         if (!ModelState.IsValid) return BadRequest(ModelState);
-        return await _tagService.GetAllTags();
+        return await _tagService.GetTags();
     }
     
     
@@ -43,9 +43,43 @@ public class TagsController : ApiController
      * Deletes a tag by ID
      */
     [HttpDelete("{tagId}")]
-    public async Task<IActionResult> DeleteTag(string tagId)
+    public async Task<IActionResult> DeleteTag(DeleteTagRequest requestParams)
     {
         if (!ModelState.IsValid) return BadRequest(ModelState);
-        return await _tagService.DeleteTag(tagId);
+        return await _tagService.DeleteTag(requestParams);
+    }
+    
+    [HttpGet("{imageId}/get")]
+    public async Task<IActionResult> GetImageTag(string imageId)
+    {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
+        return await _tagService.GetImageTag(imageId);
+    }
+    
+    [HttpGet("{imageId}/getexcluded")]
+    public async Task<IActionResult> GetTagsNotOnImage(string imageId)
+    {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
+        return await _tagService.GetTagsNotOnImage(imageId);
+    }
+
+    [HttpPost("{imageId}/add")]
+    public async Task<IActionResult> AddTagsToImage(string imageId, string tagId)
+    {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
+        return await _tagService.AddTagToImage(imageId, tagId);
+    }
+
+    [HttpDelete("{imageId}/deletefromimage")]
+    public async Task<IActionResult> RemoveTagsFromImage(string imageId, string tagId)
+    {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
+        return await _tagService.RemoveTagFromImage(imageId, tagId);
+    }
+
+    public async Task<IActionResult> GetAssetsByTag(GetAssetsByTagsRequest requestParams)
+    {
+        if (!ModelState.IsValid) return BadRequest(ModelState);
+        return await _tagService.GetAssetsByTags(requestParams);
     }
 }
